fix : 카테고리 선택시 튕기는 문제 해결
nullable하게 받아오는 데이터 타입 설정
튜터님들 사용하실 수 있게 API키 코드 내에 넣음

+++ app/src/main/java/com/example/meohaji/home/CategoryChannel.kt
@@ -9,8 +9,8 @@ data class CategoryChannel(
     val title: String,
     val thumbnail: String,
     val description: String,
-    val viewCount: Int,
-    val subscriberCount: Int,
-    val videoCount: Int,
+    val viewCount: Long,
+    val subscriberCount: Long,
+    val videoCount: Long,
     val customUrl: String
 ): Parcelable
